Restructures a fair amount of code.
The default variables for an object are now defined in its Create event,
rather than the Step event. Globals that were defined in the Game object
are now moved to relevant objects. This may change later - variables set
in the Create event can't be accessed unless an instance of that object
currently exists (in room?), and some cases may call for accessing them at
other times.